Show draftjs, md and blocks side by side
